Update NFSP syntax to require a speaker and subject for each block
The previous syntax specified the speaker on the outside of the block
and the subject on individual statements in the block. The new syntax
specifies the speaker and subject outside the block, and leaves the
subject implicit for statements inside the block. This avoids situations
where the speaker could try to address themself, which is not valid
Shakespeare.