Update NFSP syntax, possum parser, possum AST
The new syntax iteration brings back scopes for speaker variables. I
briefly thought that specifying the subject var was enough, but the
speaker needs to be explicit in order to generate Shakespeare code. We
use `foo.{ }` instead of `foo { }` to emphasize that `foo` must be
staged, similar to `bar.set(0)`.